# docs/openapi-skeleton.yaml
openapi: 3.0.3
info:
  title: Veridrop
  version: 0.1.0
  description: 'Proofed Curation Market control plane: stake drops, verify delivery proofs, allocate rewards.'
  x-product-slug: ai-decentralized-veridrop
  x-source: ai-in-decentralized+ai/openminedOcean Protocol Technical Whitepaper/
servers:
- url: https://api.example.com
...
paths:
...
  /v1/reward-epochs:
    get:
      operationId: listRewardEpochs
      tags:
      - RewardEpochs
      summary: List RewardEpoch
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      parameters:
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelopeRewardEpoch'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: openRewardEpoch
      tags:
      - RewardEpochs
      summary: Create RewardEpoch
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RewardEpochCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelopeRewardEpoch'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/reward-epochs/{epochId}:
    parameters:
    - $ref: '#/components/parameters/EpochId'
    get:
      operationId: getRewardEpoch
      tags:
      - RewardEpochs
      summary: Get RewardEpoch
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelopeRewardEpoch'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/arbitration-cases:
    get:
      operationId: listArbitrationCases
      tags:
      - ArbitrationCases
      summary: List ArbitrationCase
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      parameters:
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelopeArbitrationCase'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: openArbitrationCase
      tags:
      - ArbitrationCases
      summary: Create ArbitrationCase
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArbitrationCaseCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelopeArbitrationCase'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/arbitration-cases/{caseId}:
    parameters:
    - $ref: '#/components/parameters/CaseId'
    get:
      operationId: getArbitrationCase
      tags:
      - ArbitrationCases
      summary: Get ArbitrationCase
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelopeArbitrationCase'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    EpochId:
      name: epochId
      in: path
      required: true
      schema:
        type: string
    CaseId:
      name: caseId
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    ResponseMeta:
      type: object
      required:
      - requestId
      - timestamp
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    Problem:
      type: object
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
...
    RewardEpoch:
      type: object
      required:
      - id
      - periodStart
      - periodEnd
      - status
      - totalEmitted
      properties:
        id:
          type: string
        periodStart:
          type: string
          format: date-time
        periodEnd:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - open
          - finalizing
          - paid
          - paused
        totalEmitted:
          type: number
    RewardEpochCreate:
      type: object
      required:
      - periodStart
      - periodEnd
      properties:
        periodStart:
          type: string
          format: date-time
        periodEnd:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - open
          - finalizing
          - paid
          - paused
        totalEmitted:
          type: number
    DataEnvelopeRewardEpoch:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          $ref: '#/components/schemas/RewardEpoch'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ListEnvelopeRewardEpoch:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          type: object
          required:
          - items
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/RewardEpoch'
            nextCursor:
              type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ArbitrationCase:
      type: object
      required:
      - id
      - subjectType
      - subjectId
      - status
      - claim
      properties:
        id:
          type: string
        subjectType:
          type: string
          enum:
          - service
          - agreement
          - provenance
        subjectId:
          type: string
        status:
          type: string
          enum:
          - open
          - resolved
          - dismissed
        claim:
          type: string
        resolution:
          type: string
    ArbitrationCaseCreate:
      type: object
      required:
      - subjectType
      - subjectId
      - claim
      properties:
        subjectType:
          type: string
          enum:
          - service
          - agreement
          - provenance
        subjectId:
          type: string
        status:
          type: string
          enum:
          - open
          - resolved
          - dismissed
        claim:
          type: string
        resolution:
          type: string
    DataEnvelopeArbitrationCase:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          $ref: '#/components/schemas/ArbitrationCase'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ListEnvelopeArbitrationCase:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          type: object
          required:
          - items
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ArbitrationCase'
            nextCursor:
              type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'