# platform/api-server/src/domains/arbitration/handlers/arbitration-case.handlers.ts
/**
 * Handlers for ArbitrationCase
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: arbitration
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { ArbitrationDomainModule } from "../dependencies/arbitration-ddd.dependencies.js";
import type { GetArbitrationCaseInput, ListArbitrationCasesInput, OpenArbitrationCaseInput } from "@veridrop/services/arbitration";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v0/arbitration-cases
 * Operation: listArbitrationCases
 */
export async function listArbitrationCases(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], status?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: ArbitrationDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListArbitrationCasesInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.cases.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v0/arbitration-cases
 * Operation: openArbitrationCase
 */
export async function openArbitrationCase(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: ArbitrationDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: OpenArbitrationCaseInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.cases.get.execute(input);
  return reply.code(201).send(result);
}


/**
 * Handler for GET /v0/arbitration-cases/{caseId}
 * Operation: getArbitrationCase
 */
export async function getArbitrationCase(
  request: FastifyRequest<{ Params: { caseId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: ArbitrationDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetArbitrationCaseInput = {
    caseId: request.params.caseId,
  };
  const result = await deps.useCases.cases.get.execute(input);
  return reply.code(200).send(result);
}