# platform/api-server/src/domains/rewards/handlers/reward-epoch.handlers.ts
/**
 * Handlers for RewardEpoch
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: rewards
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { RewardsDomainModule } from "../dependencies/rewards-ddd.dependencies.js";
import type { GetRewardEpochInput, ListRewardEpochsInput, OpenRewardEpochInput } from "@veridrop/services/rewards";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v0/reward-epochs
 * Operation: listRewardEpochs
 */
export async function listRewardEpochs(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: RewardsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListRewardEpochsInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.rewardEpoches.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v0/reward-epochs
 * Operation: openRewardEpoch
 */
export async function openRewardEpoch(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: RewardsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: OpenRewardEpochInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.rewardEpoches.get.execute(input);
  return reply.code(201).send(result);
}


/**
 * Handler for GET /v0/reward-epochs/{epochId}
 * Operation: getRewardEpoch
 */
export async function getRewardEpoch(
  request: FastifyRequest<{ Params: { epochId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: RewardsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetRewardEpochInput = {
    epochId: request.params.epochId,
  };
  const result = await deps.useCases.rewardEpoches.get.execute(input);
  return reply.code(200).send(result);
}